requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
python-dotenv
flask_cors
flask_wtf
gunicorn
requests
Faker